renamed Landing.js to Landing.jsx
